Refactor models, add comment to magic code
