+++ client-app/src/components/Error.vue
@@ -33,8 +33,8 @@ import Vue from "vue";
 import { Component, Prop } from "vue-property-decorator";
 import { faCaretRight, faCaretDown } from '@fortawesome/free-solid-svg-icons'
 import { Subscription } from "rxjs";
-import { ReactiveErrorInfo } from "@common/models/error-info.model";
 import { REMOVE_ERROR } from "@init-app/store/definitions";
+import { ReactiveErrorInfo } from "@init-app/store/types";
 
 @Component({
   name: "InitApp"
