+++ client-app/src/pages/init/App.vue
@@ -14,9 +14,9 @@ import Vue from "vue";
 import Component from "vue-class-component";
 import { faCaretRight, faCaretDown } from '@fortawesome/free-solid-svg-icons'
 import { Subscription } from "rxjs";
-import { ReactiveErrorInfo } from "@common/models/error-info.model";
 import Error from "@components/Error.vue";
 import { REMOVE_ERROR } from "@init-app/store/definitions";
+import { ReactiveErrorInfo } from "@init-app/store/types";
 
 @Component({
   name: "InitApp",
