+++ client-app/src/pages/init/main.ts
@@ -1,10 +1,10 @@
 import Vue from "vue";
 import i18n from "@i18n";
 import "styles/default.scss";
-import ErrorInfo from '@common/models/error-info.model';
 import InitializationService from '@common/services/initialization.service';
 import App from "@init-app/App.vue";
 import store from "@init-app/store";
+import ErrorInfo from '@init-app/store/types';
 import { ADD_ERROR } from './store/definitions';
 
 window.onerror = (message, source, line, column, error) => {
