+++ client-app/src/pages/init/store/mutations.ts
@@ -1,7 +1,7 @@
 import { MutationTree } from "vuex";
-import ErrorInfo from '@common/models/error-info.model';
 import { ADD_ERROR, REMOVE_ERROR } from '@init-app/store/definitions';
 import { State } from "@init-app/store/state";
+import ErrorInfo from '@init-app/store/types';
 
 //mutations
 export const mutations: MutationTree<State> = {
