+++ client-app/src/pages/init/store/types.ts
