+++ templates/customers/login.liquid
@@ -23,9 +23,9 @@
                 <div class="custom-control custom-checkbox mt-3">
                     <input id="RememberMe" class="custom-control-input" type="checkbox">
                     <label for="RememberMe" class="custom-control-label">Remember me</label>
-                </div> 
+                </div>
                 <button type="submit" class="btn btn-primary mt-2">Log in</button>
-                <a class="forgot-link align-self-end mt-3" href="{{ '/account/forgotpassword' | absolute_url }}">Forgot the password?</a>
+                <a class="mt-3" href="{{ '/account/forgotpassword' | absolute_url }}">Forgot the password?</a>
                 {{ antiforgery }}
             </form>
         </div>
