+++ templates/index.liquid
@@ -1,3 +1,4 @@
-<div class="d-flex flex-grow-1 justify-content-center mt-3">
-  <button class="btn btn-danger" onclick="throw new Error();">Throw error from liquid</button>
-</div>
+<button class="btn btn-danger my-3"
+        onclick="throw new Error();">
+  Throw error from liquid
+</button>
